% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUwieAvg.R
\name{AICweights}
\alias{AICweights}
\title{Calculate AIC weights}
\usage{
AICweights(AIC, na.rm = TRUE)
}
\arguments{
\item{AIC}{A vector of values.}

\item{na.rm}{Whether to remove NA values.}
}
\value{
A named vector of weights with names inherited from \code{AIC}.
}
\description{
This function takes a vector of AIC (Akaike Information Criterion) values and returns a vector of AIC weights using the formula from Burnham and Anderson (2002).
}
\details{
If \code{remove.na = FALSE} and any values in \code{AIC} are \code{NA}, all returned values will be \code{NA}.
}
\examples{
AIC <- c(NA, 5, 10, 20, 25)
#ignore NAs
AICweights(AIC)

#should return all NAs
AICweights(AIC, na.rm = FALSE)
}
