% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUwieAvg.R
\name{deltaAIC}
\alias{deltaAIC}
\title{Calculate deltaAIC}
\usage{
deltaAIC(AIC, na.rm = TRUE)
}
\arguments{
\item{AIC}{A vector of values.}

\item{na.rm}{Whether to remove NA values.}
}
\value{
A vector of weights.
}
\description{
Calculate deltaAIC (Akaike Information Criterion), the absolute difference between the lowest AIC value and the other AIC values.
}
\details{
If \code{remove.na = FALSE} and any values in \code{AIC} are \code{NA}, all returned values will be \code{NA}.
}
\examples{
AIC <- c(NA, 5, 10, 20, 25)
deltaAIC(AIC)

#should return all NAs
deltaAIC(AIC, na.rm = FALSE)
}
