% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSimmapDensity.R
\name{extractSimmapDensity}
\alias{extractSimmapDensity}
\title{Extract posterior values from densityMap and/or contMap objects}
\usage{
extractSimmapDensity(...)
}
\arguments{
\item{...}{A vector of values.}
}
\value{
A data.frame where rows represent the mapped traits through time and columns represent each of the supplied objects in \code{...}. The column names are set to the names of objects. Also includes a column for the edge index and a column for the age (in the units of the tree).
}
\description{
This function takes a list of densityMap and/or contMap objects and extracts and aggregates the posterior values across them. The age and edge index of each posterior value are also included.
}
\details{
Assumes that the gradient resolution is identical for all objects (specified with \code{res} in \code{densityMap} and \code{contMap}). If the tree has a \code{root.time} element, this will be used as the max age, otherwise \code{max(nodeHeights(tree))} will be used.
}
\examples{
library(phytools)
# simulate tree
tree <- pbtree(n=70,scale=1)

# simulate discrete trait
Q <- matrix(c(-1,1,1,-1),2,2)
rownames(Q) <- colnames(Q)<-c(0,1)
x1 <- sim.history(tree,Q)$states

# generate stochastic maps and density map
mtrees <- make.simmap(tree,x1,nsim=100)
map1 <- densityMap(mtrees)

# simulate continuous trait
x2 <- fastBM(tree,sig2=0.1)

# generate cont map
map2 <- contMap(tree, x2)

# extract posterior densities
pp_data <- extractSimmapDensity(map1, map2)

# plot to see how traits have evolved with respect to one another
plot(pp_data$map1, pp_data$map2)
}
