% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMappedStates2.R
\name{mergeMappedStates2}
\alias{mergeMappedStates2}
\title{Rename, merge, or split mapped states}
\usage{
mergeMappedStates2(tree, old.states, new.state, node = NULL)
}
\arguments{
\item{tree}{an object of class "simmap" or "multiSimmap" containing one or more phylogenetic trees with a mapped discrete character}

\item{old.states}{state(s) to rename or merge}

\item{new.state}{name for new state}

\item{node}{a node index, used to specify a subclade of interest}
}
\value{
An object of class "simmap" or "multiSimmap".
}
\description{
This function renames, merges, or splits mapped states on a tree.
}
\details{
If \code{node} is specified, only mapped states of descendants edges of that node are modified. This function is based on the similarly named utility function by Liam Revell in \code{phytools}.
}
\examples{
library(phytools)
# simulate a mapped tree
set.seed(4)
Q <- matrix(c(-2,1,1,1,-2,1,1,1,-2),3,3)
rownames(Q) <- colnames(Q) <- letters[1:3]
tree <- sim.history(pbtree(n=100,scale=1),Q)
cols <- setNames(c("blue","red","green","orange"),letters[1:4])

# plot the mapping
plot(tree, cols, ftype="i", fsize=0.7)

# split state c to state d within subclade
tree2 <- mergeMappedStates2(tree, "c", "d", 173)

# plot the new mapping
plot(tree2, cols, ftype="i", fsize=0.7)
}
