% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUwieAvg.R
\name{OUwieAICSumm}
\alias{OUwieAICSumm}
\title{Summarize the model fit across (many) OUwie results using AICc}
\usage{
OUwieAICSumm(ou.parameters, na.rm = TRUE)
}
\arguments{
\item{ou.parameters}{An object of class \code{ouwiepars} as output from \code{OUwieParSumm} or \code{CleanOUwieParameters}}

\item{na.rm}{Whether to ignore model results with \code{NA} AICc values}
}
\value{
A list with two elements:
\item{Weights}{A data.frame of the AICc weights for each model across the replicates}
\item{Counts}{A named vector giving the number of replicates in which each model has the highest AICc weight}
}
\description{
Returns AICc weights and counts for OUwie results that have been processed using \code{OUwieParSumm}.
}
\details{
\code{na.rm = TRUE} will remove any models where \code{AICc == NA}, but will use other models for AIC weight calculation;
\code{na.rm = FALSE} will remove any replicates (rows of the output) with any models where \code{AICc == NA}.
}
\examples{
ou.parameters <- OUwieParSumm(ou.results, regime.mat)
OUwieAICSumm(ou.parameters)
}
