% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUwieAvg.R
\name{OUwieCleanPar}
\alias{OUwieCleanPar}
\title{Clean extracted parameters from OUwie results}
\usage{
OUwieCleanPar(ou.parameters, lower = list(), upper = list())
}
\arguments{
\item{ou.parameters}{An object of class \code{ouwiepars} as output from \code{OUwieParSumm}}

\item{lower}{A list of lower bounds for model parameters}

\item{upper}{A list of upper bounds for model parameters}
}
\value{
An \code{ouwiepars} object
}
\description{
Cleans the parameters that are extracted by \code{OUwieParamSum} based on user-specified upper and lower bounds.
}
\details{
Parameter estimates outside of these bounds will result in the AICc being changed to NA,
which will affect downstream model averaging (see \code{OUwieModelAvg}).
}
\examples{
ou.parameters <- OUwieParSumm(ou.results, regime.mat)

#Sets the AICc for the BM1 model to NA, so it wouldn't be included in downstream model averaging
OUwieCleanPar(ou.parameters, upper = list("AICc" = 45))
}
