% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUwieAvg.R
\name{OUwieModelAvg}
\alias{OUwieModelAvg}
\title{Model average the parameters across (many) OUwie results using AICc}
\usage{
OUwieModelAvg(ou.parameters, OU.only = FALSE, na.rm = TRUE)
}
\arguments{
\item{ou.parameters}{An object of class \code{ouwiepars} as output from \code{OUwieParSumm} or \code{CleanOUwieParameters}}

\item{OU.only}{Whether any Brownian motion models should be dropped before model averaging}

\item{na.rm}{Whether to ignore model results with \code{NA} AICc values}
}
\value{
A list with two elements:
\item{Weights}{A data.frame of the AICc weights for each model across the replicates}
\item{Counts}{A named vector giving the number of replicates in which each model has the highest AICc weight}
}
\description{
Internally calculates AICc weights and uses them to model average the parameter values output from \code{OUwieParSumm}.
}
\details{
\code{na.rm = TRUE} will remove any models where \code{AICc == NA}, but will use other models for model averaging;
\code{na.rm = FALSE} will remove any replicates (rows of the output) with any models where \code{AICc == NA}.
}
\examples{
ou.parameters <- OUwieParSumm(ou.results, regime.mat)
OUwieModelAvg(ou.parameters)
}
